Update AWAP section
Ivan removed the RODBC dependency, so it's now much easier to install
awaptools from github and use that to fetch the grid files.

In this commit, the functions are removed, awaptools are installed from
GitHub and more detail is added about AWAP.

I also corrected capitalisation of CRAN and GitHub in one other spot
where I saw it.